+++ pricing/src/epilot/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.44.0'
-    gen_version: str = '2.150.0'
-    user_agent: str = 'speakeasy-sdk/python 1.44.0 2.150.0 1.0.0 epilot-pricing'
+    sdk_version: str = '1.44.1'
+    gen_version: str = '2.154.1'
+    user_agent: str = 'speakeasy-sdk/python 1.44.1 2.154.1 1.0.0 epilot-pricing'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: